Add type to types builtins
As it is builtins is not part of the typing. I couldn't understand where the declaration happen. Doesn't export const types: typeof pgTypes; should handle it ???

The above is a fix. that maybe not liked. If there is a better way. Please show me the  way.

# Context:
    "@types/pg": "^7.4.14",
    "pg": "^7.11.0",
    "pg-types": "^2.1.0"

pg.types. => doesn't show builtins at all 
but only
-> 
getTypeParser
setTypeParser
arrayParser